Remove iMazing suggestion for getting user data from iOS
Two people have tried, and user data was not present in either case.
There were no cached resources or masters, either, and since on Android
user data is located together with those, I am going to assume that on
iOS all of this is stored in some other place that iMazing can't access.