Disable android:debuggable in APKs to fix crash/freeze on GrapheneOS
No idea why it happens, but this is the one change undoing which fixes it.

Unrelated to the crash, the new builds also restore the
original AndroidManifest's android:compileSdkVersion="30"
android:compileSdkVersionCodename="11", which seem to be
informative only. I believe I previously removed them because
apktool refused to build APKs when they were retained, but
perhaps I've updated apktool since then, because it works now.